Add files via upload
This dataset is referenced in the IDN CP Guidance doc but is missing from the catalogue. It's not clear which catalogue it belongs to (i.e. ISU or ANUFP) so adding it to the demo catalogue. This is an exceptional case, just trying to validate the guidance example without having to make a new one.